+++ lib/page/home/search_widget.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class SearchWidget extends StatelessWidget {
+  const SearchWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 16.0),
+      child: TextField(
+        decoration: InputDecoration(
+            prefixIcon: const Icon(Icons.search),
+            hintText: 'Search',
+            border: OutlineInputBorder(
+              borderRadius: BorderRadius.circular(24),
+            ),
+            contentPadding: const EdgeInsets.symmetric(horizontal: 16)),
+      ),
+    );
+  }
+}
